Simplify example OIDC configuration for Authelia
This may be a controversial change since the example configuration is no longer a full working example. However, the existing example was not secure and a user copy/pasting the example without paying attention would end up with an insecure configuration. Someone setting this up should be forced to go to the Authelia documentation and read through what all the settings mean.

In addition, the settings I have removed from the example are all either unnecessary to specify, or should be specified using a secret instead of an inline value. My own personal working configuration actually does match the new example as shown, because the other settings that are required to be specified are done so via secrets.